Added some documentation and fixed Fourier mode
I was not using Fourier mode the way I thought I had designed it... I was updating the component number instead of the number of modes. Updating the modes is a bit more involved but I just modify the default value by the mode number and then let the music modify the result.